requirements.txt
mem0ai>=0.1.20
openai>=1.50.0
chromadb>=0.5.5
pandas>=2.2.0
faker>=30.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
